Taskの型を修正したのでControllerとRepositoryも修正

+++ springboot-demo/src/main/kotlin/todolist/TaskRepository.kt
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface TaskRepository : JpaRepository<Task, Long>
+interface TaskRepository : JpaRepository<Task, Int>
